Move styling to a brand new `iced_style` crate
