+++ style/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "iced_style"
+version = "0.1.0-alpha"
+authors = ["Héctor Ramón Jiménez <hector0193@gmail.com>"]
+edition = "2018"
+description = "The default set of styles of Iced"
+license = "MIT"
+repository = "https://github.com/hecrj/iced"
+documentation = "https://docs.rs/iced_style"
+keywords = ["gui", "ui", "graphics", "interface", "widgets"]
+categories = ["gui"]
+
+[dependencies]
+iced_core = { version = "0.1.0", path = "../core" }
