+++ style/src/button.rs
@@ -1,0 +1,79 @@
+//! Allow your users to perform actions by pressing a button.
+use iced_core::{Background, Color};
+
+/// The appearance of a button.
+#[derive(Debug)]
+pub struct Style {
+    pub shadow_offset: f32,
+    pub background: Option<Background>,
+    pub border_radius: u16,
+    pub text_color: Color,
+}
+
+/// A set of rules that dictate the style of a button.
+pub trait StyleSheet {
+    fn active(&self) -> Style;
+
+    fn hovered(&self) -> Style {
+        let active = self.active();
+
+        Style {
+            shadow_offset: active.shadow_offset + 1.0,
+            ..active
+        }
+    }
+
+    fn pressed(&self) -> Style {
+        Style {
+            shadow_offset: 0.0,
+            ..self.active()
+        }
+    }
+
+    fn disabled(&self) -> Style {
+        let active = self.active();
+
+        Style {
+            shadow_offset: 0.0,
+            background: active.background.map(|background| match background {
+                Background::Color(color) => Background::Color(Color {
+                    a: color.a * 0.5,
+                    ..color
+                }),
+            }),
+            text_color: Color {
+                a: active.text_color.a * 0.5,
+                ..active.text_color
+            },
+            ..active
+        }
+    }
+}
+
+struct Default;
+
+impl StyleSheet for Default {
+    fn active(&self) -> Style {
+        Style {
+            shadow_offset: 1.0,
+            background: Some(Background::Color([0.5, 0.5, 0.5].into())),
+            border_radius: 5,
+            text_color: Color::BLACK,
+        }
+    }
+}
+
+impl std::default::Default for Box<dyn StyleSheet> {
+    fn default() -> Self {
+        Box::new(Default)
+    }
+}
+
+impl<T> From<T> for Box<dyn StyleSheet>
+where
+    T: 'static + StyleSheet,
+{
+    fn from(style: T) -> Self {
+        Box::new(style)
+    }
+}
